requirements.txt
requests >= 2.30.0
pydantic >= 2.3.0
pydantic-settings >= 2.0.3
pymongo >= 4.5.0
fastapi >= 0.103.2
uvicorn >= 0.23.2
python-multipart >= 0.0.6
PyGithub >= 2.5.0
gitpython >= 3.1.43
loguru>=0.7.2
redis>=5.0.1
joblib>=1.3.2
rq>=1.16.1
setproctitle>=1.3.3
neo4j>=5.18.0
semver>=3.0.2
packaging>=24.0
johnnydep>=1.20.4
matplotlib>=3.7.5
crispyn==0.0.6
docker==7.1.0
pandas>=2.0.3
igraph>=0.11.6
networkx>=3.1
tqdm>=4.67.0